+++ OpenGL_Renderer/src/Engine/Shader_PBR_Uniforms.h
@@ -101,12 +101,12 @@ namespace uniform {
 			ImGui::Checkbox("cast Shadows", &castShadows);
 			ImGui::Checkbox("recieve Shadows", &recieveShadows);
 
-			for (size_t i = 0; i < numberOfTextures; i++) {
+			for (int i = 0; i < numberOfTextures; i++) {
 				std::string name;
 				name = "tex : " + std::to_string(i);
 				if (ImGui::CollapsingHeader(name.c_str(), ImGuiTreeNodeFlags_DefaultOpen | ImGuiTreeNodeFlags_AllowItemOverlap)) {
-					int adjustedWidth = ImGui::GetContentRegionAvailWidth() * 1024 / 1024;
-					ImGui::Image((void*)_pbr_textures[i].GetTexture(), ImVec2(ImGui::GetContentRegionAvailWidth(), adjustedWidth));
+					int adjustedWidth = (int)ImGui::GetContentRegionAvailWidth() * 1024 / 1024;
+					ImGui::Image((void*)_pbr_textures[i].GetTexture(), ImVec2(ImGui::GetContentRegionAvailWidth(), (int)adjustedWidth));
 				}
 			}
 			ImGui::Unindent();
@@ -119,7 +119,7 @@ namespace uniform {
 		numberOfTextures = tex.size();
 		_pbr_textures = std::make_unique<Texture2D[]>(numberOfTextures);
 
-		for (size_t i = 0; i < numberOfTextures; i++){
+		for (int i = 0; i < numberOfTextures; i++){
 			_pbr_textures[i].LoadTexture(tex[i], true);	  
 		}
 	}
